Modify Calculator::main to utilize new input handling.
Now recognizes commands: 'avg', 'sum', 'exit'.
avg -> prints average of previous results in console
sum -> prints sum of previous results in console
exit -> stops the app
Any other input results in the app proceeding with a calculation.
Still need to fix crashing when input is not a calculation and none
of the above.